Add files via upload
The dataset contains the smoothed functional observations of the IAGOS data. This RData object is specifically used to execute the script "fGGRM IAGOS DATA.R". Details about the smoothing procedure can be found in the manuscript under the section "Evaluation of the Scores" or in the accompanying script "IAGOS data import.R".